Add files via upload
This project uses a Convolutional Neural Network (CNN) to recognize handwritten mathematical symbols. The images are preprocessed (resized and normalized), and the CNN model is trained on over 80 symbol classes. The model learns to extract features and classify symbols accurately, even when they are visually similar. Performance is evaluated using accuracy, loss, and a confusion matrix.